project_root/python_server/requirements.txt
flask-cors
torch
torchvision
ultralytics
pillow
opencv-python